+++ src/DeleteDocumentManager.php
@@ -12,7 +12,7 @@ class DeleteDocumentManager extends AbstractManager
 
     protected string $identifier;
 
-    public function makeRequest(): bool
+    public function makeRequest(): array
     {
         try {
             /** @var \GuzzleHttp\Client */
@@ -22,12 +22,12 @@ class DeleteDocumentManager extends AbstractManager
                 $this->getCompleteUrl(),
                 [
                     'json' => [
-                        'identifier' => $this->getIdentifier()
-                    ]
+                        'identifier' => $this->getIdentifier(),
+                    ],
                 ]
             );
         } catch (ServerException $th) {
-            if (!$th->hasResponse()) {
+            if (! $th->hasResponse()) {
                 throw new InvalidResponseException("Generic server error");
             }
 
@@ -42,14 +42,14 @@ class DeleteDocumentManager extends AbstractManager
 
         $status = json_decode($response->getBody()->getContents(), true)['status'] ?? false;
 
-        if (!$status) {
+        if (! $status) {
             throw new CannotDeleteFileException(json_decode($response->getBody()->getContents(), true)['error'] ?? "Cannot delete file");
         }
 
-        return true;
+        return ["status" => true];
     }
 
-    public function delete(): bool
+    public function delete(): array
     {
         return $this->makeRequest();
     }