+++ tests/TextUploadManagerTest.php
@@ -28,7 +28,7 @@ class TextUploadManagerTest extends TestCase
     {
 
         $mock = new MockHandler([
-            new Response($status_code, [], json_encode($response_message))
+            new Response($status_code, [], json_encode($response_message)),
         ]);
 
         $handlerStack = HandlerStack::create($mock);
@@ -36,7 +36,7 @@ class TextUploadManagerTest extends TestCase
 
         $manager = Manager::textUpload()->setClient($clientMock);
 
-        if($thows) {
+        if ($thows) {
             $this->expectException($thows);
         }
 
@@ -48,10 +48,11 @@ class TextUploadManagerTest extends TestCase
             ->setLink($data["link"])
             ->setPermissions($data["permissions"])
             ->setLanguage($data["language"])
+            ->setMetadata($data["metadata"] ?? [])
             ->setDocumentDate($data["document_date"] ?? "")
             ->upload();
 
-        $this->assertTrue($status);
+        $this->assertEquals(['status' => true], $status);
 
     }
 
@@ -76,13 +77,14 @@ Illius oritur primum quanto vacuitate? Has loco pácem quisquam, variis vétéru
                     "link" => "https://www.example.com",
                     "language" => "it",
                     "permissions" => ["a", "b", "c"],
-                    "document_date" => "2020-01-01"
+                    "metadata" => ["a" => "b", "b" => "c", "c" => "d"],
+                    "document_date" => "2020-01-01",
                 ],
                 "status_code" => 200,
                 "response_message" => [
-                    "status" => true
+                    "status" => true,
                 ],
-                "throws" => null
+                "throws" => null,
             ],
             "with invalid payload" => [
                 "data" => [
@@ -97,9 +99,9 @@ Illius oritur primum quanto vacuitate? Has loco pácem quisquam, variis vétéru
                 ],
                 "status_code" => 200,
                 "response_message" => [
-                    "status" => true
+                    "status" => true,
                 ],
-                "throws" => DataInvalidException::class
+                "throws" => DataInvalidException::class,
             ],
             "with http error" => [
                 "data" => [
@@ -110,15 +112,14 @@ Illius oritur primum quanto vacuitate? Has loco pácem quisquam, variis vétéru
                     "link" => "https://www.example.com",
                     "language" => "it",
                     "permissions" => ["a", "b", "c"],
-                    "document_date" => "2020-01-01"
+                    "document_date" => "2020-01-01",
                 ],
                 "status_code" => 500,
                 "response_message" => [
-                    "error" => "Some error"
+                    "error" => "Some error",
                 ],
-                "throws" => InvalidResponseException::class
+                "throws" => InvalidResponseException::class,
             ],
         ];
     }
-
 }